requirements.txt
opencv-python>=4.10.0.84
mediapipe>=0.10.14
numpy>=1.26.0
python-dotenv>=1.0.1
google-generativeai>=0.8.3
openai>=1.44.0